depricate a rkk receiver and refactor a rkk remote
