+++ Plugins/Smartsuit/Source/Smartsuit/Private/VirtualProductionSource.cpp
@@ -17,7 +17,7 @@
 #include "Runtime/JsonUtilities/Public/JsonObjectConverter.h"
 #include "Serialization/BufferArchive.h"
 #include "lz4frame.h"
-#include "RokokoReceiver.h"
+
 
 TSharedPtr<FVirtualProductionSource> FVirtualProductionSource::instance = nullptr;
